backend/user_service/requirements.txt
# backend/user_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib[bcrypt]
bcrypt
python-jose[cryptography]
email-validator

python-multipart # <--- TAMBAHKAN BARIS INI